+++ core/components/crosslinks/lexicon/de/default.inc.php
@@ -2,20 +2,21 @@
 $_lang['crosslinks'] = 'Crosslinks';
 
 $_lang['crosslinks.management'] = 'Crosslinks-Manager';
+
 $_lang['crosslinks.menu'] = 'Crosslinks';
 $_lang['crosslinks.menu_desc'] = 'Crosslinks verwalten';
 
 $_lang['crosslinks.links'] = 'Crosslinks';
 $_lang['crosslinks.links_desc'] = 'Verwalten Sie Ihre Crosslinks hier. Sie können eine Tabellenzeile mit den Icons oder mit einem Rechtsklick bearbeiten und löschen.';
 
-$_lang['crosslinks.link'] = 'Link';
 $_lang['crosslinks.link_create'] = 'Link hinzufügen';
+$_lang['crosslinks.link_duplicate'] = 'Link duplizieren';
+$_lang['crosslinks.link_parameter'] = 'Parameter';
 $_lang['crosslinks.link_remove'] = 'Link entfernen';
 $_lang['crosslinks.link_remove_confirm'] = 'Sind Sie sicher, dass Sie diesen Link entfernen möchten?';
 $_lang['crosslinks.link_resource'] = 'Ressource';
 $_lang['crosslinks.link_text'] = 'Text';
 $_lang['crosslinks.link_update'] = 'Link bearbeiten';
-$_lang['crosslinks.link_parameter'] = 'Parameter';
 
 $_lang['crosslinks.jsongrid_key'] = 'Schlüssel';
 $_lang['crosslinks.jsongrid_value'] = 'Wert';
@@ -25,11 +26,12 @@ $_lang['crosslinks.link_err_ns_resource'] = 'Sie müssen eine Ressource auswähl
 $_lang['crosslinks.link_err_ns_text'] = 'Sie müssen einen Text eingeben';
 $_lang['crosslinks.link_err_nv_text'] = 'Der Text enthält ungültige Zeichen, Sie können Buchstaben, Zahlem, Leerzeichen und folgende Zeichen benutzen: :.;,-_';
 
-$_lang['setting_crosslinks.debug'] = 'Debug';
-$_lang['setting_crosslinks.debug_desc'] = 'Debug-Informationen im MODX Fehlerprotokoll ausgeben.';
-$_lang['setting_crosslinks.fullwords'] = 'Suche Ganze Wörter';
-$_lang['setting_crosslinks.fullwords_desc'] = 'Nur ganze Worte eines Crosslinksbegriffs im Ressourcen-Inhalt ersetzen.';
-$_lang['setting_crosslinks.sections'] = 'Nur in Abschnitten';
-$_lang['setting_crosslinks.sections_desc'] = 'Crosslinks Begriffe nur in mit &lt;!— CrosslinksStart --&gt; und &lt;!— CrosslinksEnd --&gt; markierten Abschnitten ersetzen.';
-$_lang['setting_crosslinks.tpl'] = 'Highlight Template';
-$_lang['setting_crosslinks.tpl_desc'] = 'Template Chunk für das Higlight Element.';
+$_lang['crosslinks.settings'] = '<i class="icon icon-cog"></i>';
+$_lang['crosslinks.settings_desc'] = 'Bearbeiten Sie die Einstellungen von Crosslinks. Sie können den Wert einer Systemeinstellung mit einem Doppelklick auf die ‚Wert‘-Tabellenzelle oder die Systemeinstellung mit einem Rechtsklick in der Tabellenzelle bearbeiten.';
+
+$_lang['crosslinks.debug_mode'] = 'Debug Modus';
+
+$_lang['crosslinks.systemsetting_key_err_nv'] = 'Sie dürfen nur Einstellungen mit dem Prefix crosslinks bearbeiten.';
+$_lang['crosslinks.systemsetting_usergroup_err_nv'] = 'Nur Mitglieder der Benutzergruppe Administrator oder Crosslinks Administrator dürfen die Einstellungen ändern.';
+
+$_lang['crosslinks.duplicate'] = 'Duplikat von';