+++ core/components/crosslinks/lexicon/de/setting.inc.php
@@ -1,0 +1,15 @@
+<?php
+$_lang['setting_crosslinks.debug'] = 'Debug';
+$_lang['setting_crosslinks.debug_desc'] = 'Debug-Informationen im MODX Fehlerprotokoll ausgeben.';
+$_lang['setting_crosslinks.disabledAttributes'] = 'Deaktivierte Attribute';
+$_lang['setting_crosslinks.disabledAttributes_desc'] = '(Kommaseparierte Liste) Crosslinks ersetzt keinen Text innerhalb von diesen HTML Tag Attributen.';
+$_lang['setting_crosslinks.fullwords'] = 'Nur ganze Worte';
+$_lang['setting_crosslinks.fullwords_desc'] = 'Nur ganze Worte eines Crosslinksbegriffs im Ressourcen-Inhalt ersetzen.';
+$_lang['setting_crosslinks.sections'] = 'Nur in Abschnitten';
+$_lang['setting_crosslinks.sections_desc'] = 'Crosslinks Begriffe nur in mit \'&lt;!— CrosslinksStart --&gt;\' und \'&lt;!— CrosslinksEnd --&gt;\' markierten Abschnitten ersetzen. Die Abschnittsmarkierungen lassen sich in den Einstellungen \'crosslinks.sectionsStart\' und \'crosslinks.sectionsEnd\' anpassen.';
+$_lang['setting_crosslinks.sectionsEnd'] = 'Abschnitt Ende Markierung';
+$_lang['setting_crosslinks.sectionsEnd_desc'] = 'Markierung am Ende eines von Crosslinks bearbeiteten Abschnitts. Das Beschränken auf markierte Abschnitte lässt sich in der Einstellung \'crosslinks.sections\' aktivieren.';
+$_lang['setting_crosslinks.sectionsStart'] = 'Abschnitt Anfang Markierung';
+$_lang['setting_crosslinks.sectionsStart_desc'] = 'Markierung am Anfang eines von Crosslinks bearbeiteten Abschnitts. Das Beschränken auf markierte Abschnitte lässt sich in der Einstellung \'crosslinks.sections\' aktivieren.';
+$_lang['setting_crosslinks.tpl'] = 'Link Template';
+$_lang['setting_crosslinks.tpl_desc'] = 'Template Chunk für das Link Element.';