+++ core/components/crosslinks/lexicon/en/setting.inc.php
@@ -1,0 +1,15 @@
+<?php
+$_lang['setting_crosslinks.debug'] = 'Debug';
+$_lang['setting_crosslinks.debug_desc'] = 'Log debug informations in MODX error log.';
+$_lang['setting_crosslinks.disabledAttributes'] = 'Disabled Attributes';
+$_lang['setting_crosslinks.disabledAttributes_desc'] = '(Comma separated list) Crosslinks does not replace text inside of this HTML tag attributes.';
+$_lang['setting_crosslinks.fullwords'] = 'Only Full Words';
+$_lang['setting_crosslinks.fullwords_desc'] = 'Replace only full words of a Crosslinks link in the resource content.';
+$_lang['setting_crosslinks.sections'] = 'Restrict to Sections';
+$_lang['setting_crosslinks.sections_desc'] = 'Replace Crosslinks links only in sections marked with \'&lt;!— CrosslinksStart --&gt;\' and \'&lt;!— CrosslinksEnd --&gt;\'. The section markers could be changed with the settings \'crosslinks.sectionsStart\' and \'crosslinks.sectionsEnd\'.';
+$_lang['setting_crosslinks.sectionsEnd'] = 'Section End Marker';
+$_lang['setting_crosslinks.sectionsEnd_desc'] = 'Marker at the end of a section processed by Crosslinks. The restriction to marked sections can be activated in the setting \'crosslinks.sections\'.';
+$_lang['setting_crosslinks.sectionsStart'] = 'Section Start Marker';
+$_lang['setting_crosslinks.sectionsStart_desc'] = 'Marker at the start of a section processed by Crosslinks. The restriction to marked sections can be activated in the setting \'crosslinks.sections\'.';
+$_lang['setting_crosslinks.tpl'] = 'Link Template';
+$_lang['setting_crosslinks.tpl_desc'] = 'Template Chunk for the link replacement.';