+++ core/components/crosslinks/docs/changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2018-12-02
+### Added
+- Don't replace crosslinks in definable html attributes
+- Edit crosslinks system settings in the custom manager page
+
 ## [1.0.0] - 2018-11-14
 ### Added
 - Initial release